Send errors and warnings as bundled messages
Now sends errors and warnings only when new errors happen or old one get removed. Sends them as one bundled message instead of lots of single messages for every error/warning. Listening for errors on the Android App is now implemented as a service, so batched messages do not get missed if the wrong tab in the app is open / app is not in foreground.